refactor: split whatsapp event from manager class
